+++ src/view/components/Variable.svelte
@@ -8,7 +8,6 @@
    const dispatch = createEventDispatcher();
 
    export let variable;
-   export let onTagClick;
    function remove() {
       dispatch("remove", variable);
    }
@@ -45,7 +44,7 @@
    </label>
 
    <Select items={argSpecs} optionIdentifier="id" labelIdentifier="id" value={variable.type} on:select={updateType} />
-   <ArgInput {onTagClick} type={variable.type} bind:value={variable.value} on:change={emitChange} />
+   <ArgInput type={variable.type} bind:value={variable.value} on:change={emitChange} />
    <button class="ui-btn ui-btn-outline ui-btn-error ui-btn-square ui-justify-self-end !ui-p-[8px]" on:click={remove}>
       <FaTimes />
    </button>